实现了新的上传逻辑 1.分享码由4位数字改为6位数字+大写字母。 2.增加了生成分享码功能（从ShareCodePool中返回）。 3.增加了下载次数限制功能，默认为一次，次数为0时文件标记为删除。 4.将FileService分离为上传/下载单独的Service。 5.在全局异常拦截器中增加了对应的异常处理。 6.更新了文件清理逻辑，每次轮询时会将过期文件暂时标记为删除以增加缓冲时间，只有已标记过的文件才会真正删除。

+++ src/main/java/ryu/cloudstoragesystem_backend/file/FileCleanupTask.java
@@ -34,15 +34,20 @@ public class FileCleanupTask {
         long now = System.currentTimeMillis();
         List<CloudFile> files = cloudFileDAO.findByTimeStampBefore(now - validTime);
         files.forEach(f -> {
-            Path path = Paths.get(fileRootPath + f.getMD5());
-            try {
-                Files.delete(path);
-                log.info("File {} delete success at {}", f.getFileName() + f.getExtension(), LocalDateTime.now());
-            } catch (IOException e) {
-                log.warn("File {} delete fail at {}, check the file root directory!", f.getFileName() + f.getExtension(), LocalDateTime.now());
+            if (!f.getRemovedFlag()){
+                f.setRemovedFlag(true);
+                cloudFileDAO.save(f);
+            }else {
+                Path path = Paths.get(fileRootPath + f.getMD5());
+                try {
+                    Files.delete(path);
+                    log.info("File {} delete success at {}", f.getFileName() + f.getExtension(), LocalDateTime.now());
+                } catch (IOException e) {
+                    log.warn("File {} delete fail at {}, check the file root directory!", f.getFileName() + f.getExtension(), LocalDateTime.now());
+                }
             }
         });
-        int count = cloudFileDAO.deleteByTimeStampBefore(now - validTime);
+        int count = cloudFileDAO.deleteByRemovedFlag(true);
         log.info("File cleanup task completed at {}, {} file(s) deleted.", LocalDateTime.now(), count);
     }
 }
